Restructure the project for new versioning approach
In issue #45, it was decided that mcproto should not end up supporting
multiple protocol versions, and instead it should only deal with the
single (latest) version.

This commit therefore restructures the project, removing all versioned
components, keeping only a single version.

This change also means a removal of the VersionMap ABC, which was used
for PACKET_MAP construction for the various game states. As this is no
longer the case, the current approach for now will be to construct the
packet maps manually, which is indeed very annoying, and a better
approach will need to be implemented later.
