+++ mcproto/packets/__init__.py
@@ -1,22 +1,13 @@
 from __future__ import annotations
 
-from mcproto.packets.interactions import (
-    PACKET_MAP,
-    async_read_packet,
-    async_write_packet,
-    sync_read_packet,
-    sync_write_packet,
-)
-from mcproto.packets.map import PacketMap
+from mcproto.packets.interactions import async_read_packet, async_write_packet, sync_read_packet, sync_write_packet
 from mcproto.packets.packet import ClientBoundPacket, GameState, Packet, PacketDirection, ServerBoundPacket
 
 __all__ = [
     "ClientBoundPacket",
     "GameState",
-    "PACKET_MAP",
     "Packet",
     "PacketDirection",
-    "PacketMap",
     "ServerBoundPacket",
     "async_read_packet",
     "async_write_packet",
