+++ mcproto/packets/interactions.py
@@ -5,11 +5,10 @@ from collections.abc import Mapping
 from typing import TypeVar
 
 from mcproto.buffer import Buffer
-from mcproto.packets.map import PacketMap
 from mcproto.packets.packet import Packet
 from mcproto.protocol.base_io import BaseAsyncReader, BaseAsyncWriter, BaseSyncReader, BaseSyncWriter
 
-__all__ = ["async_read_packet", "async_write_packet", "sync_read_packet", "sync_write_packet", "PACKET_MAP"]
+__all__ = ["async_read_packet", "async_write_packet", "sync_read_packet", "sync_write_packet"]
 
 T_Packet = TypeVar("T_Packet", bound=Packet)
 
@@ -23,8 +22,6 @@ T_Packet = TypeVar("T_Packet", bound=Packet)
 
 # Since the read functions here require PACKET_MAP, we can't move these functions
 # directly into BaseWriter/BaseReader classes, as that would be a circular import
-
-PACKET_MAP = PacketMap()
 
 
 def _serialize_packet(packet: Packet, *, compressed: bool = False) -> Buffer:
