+++ mcproto/packets/status/ping.py
