--- tests/mcproto/packets/v757/login/__init__.py
@@ -1,1 +1,0 @@
-from __future__ import annotations
